python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pillow>=9.0.0
matplotlib>=3.5.0
scikit-image>=0.19.0
tqdm>=4.64.0
ftfy>=6.1.0
regex>=2022.1.0
gradio>=4.0.0
clip @ git+https://github.com/openai/CLIP.git